bffFia/requirements.txt
anyio==4.0.0
attrs==23.1.0
black==23.10.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.1
click==8.1.7
clickclick==20.10.2
connexion==2.14.2
cryptography==41.0.4
dnspython==2.4.2
exceptiongroup==1.1.3
Flask==2.2.5
Flask-Cors==4.0.0
Flask-PyMongo==2.3.0
grpcio==1.62.0
grpcio-tools==1.62.0
gunicorn==20.1.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
inflection==0.5.1
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.19.1
jsonschema-path==0.3.1
jsonschema-specifications==2023.7.1
lazy-object-proxy==1.9.0
MarkupSafe==2.1.3
mypy-extensions==1.0.0
numpy==1.26.2
openapi-generator==1.0.6
openapi-generator-cli==4.3.1
openapi-python-generator==0.4.8
openapi-schema-pydantic==1.2.4
openapi-schema-validator==0.6.2
openapi-spec-validator==0.7.1
orjson==3.9.9
packaging==23.2
pandas==2.1.0
pathable==0.4.3
pathspec==0.11.2
platformdirs==3.11.0
protobuf==4.25.3
pycparser==2.21
pydantic==1.10.13
pymongo==4.5.0
pyOpenSSL==23.2.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986==1.5.0
rpds-py==0.10.6
six==1.16.0
sniffio==1.3.0
swagger-ui-bundle==0.0.9
tomli==2.0.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
Werkzeug==2.2.3
